requirements.txt
Flask
gunicorn
openai
python-dotenv
xai_sdk
google-auth
pyjwt
cryptography
requests
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
Flask-Cors